requirements.txt
fastapi[standard]
pydantic
psycopg[binary]==3
weasyprint
PyJWT